🔧 File heirarchy rework to please Gatsby
