+++ src/components/Contact/index.js
@@ -1,11 +1,6 @@
 import React from "react";
-import {
-  EmailIcon,
-  GitHubIcon,
-  LinkedInIcon,
-  GlobeIcon
-} from "components/icons";
-import { CONTACT_DETAILS, SOCIAL_LINKS, SITE_URL } from "constants.js";
+import { CONTACT_DETAILS, SOCIAL_LINKS, SITE_URL } from "../../constants.js";
+import { EmailIcon, GitHubIcon, LinkedInIcon, GlobeIcon } from "../icons";
 import { List, ListItem, Text, StyledLink, StyledExternalLink } from "./styles";
 
 const iconProps = {
