+++ src/components/Layout/index.js
@@ -1,0 +1,9 @@
+import React, { Fragment } from "react";
+import { GlobalStyles } from "./styles";
+
+export const Layout = ({ children }) => (
+  <Fragment>
+    <GlobalStyles />
+    {children}
+  </Fragment>
+);
