+++ src/components/Social/index.js
@@ -1,11 +1,6 @@
 import React from "react";
-import {
-  GitHubIcon,
-  MediumIcon,
-  TwitterIcon,
-  LinkedInIcon
-} from "components/icons";
-import { SOCIAL_LINKS } from "constants.js";
+import { SOCIAL_LINKS } from "../../constants.js";
+import { GitHubIcon, MediumIcon, TwitterIcon, LinkedInIcon } from "../icons";
 import { SocialLinks, SocialLink } from "./styles";
 
 const iconProps = {
